+++ src/SoftDeletes/ForceDeleteDirective.php
@@ -4,6 +4,7 @@ namespace Nuwave\Lighthouse\SoftDeletes;
 
 use GraphQL\Language\AST\FieldDefinitionNode;
 use GraphQL\Language\AST\ObjectTypeDefinitionNode;
+use Illuminate\Database\Eloquent\Builder as EloquentBuilder;
 use Illuminate\Database\Eloquent\Model;
 use Nuwave\Lighthouse\Schema\AST\DocumentAST;
 use Nuwave\Lighthouse\Schema\Directives\ModifyModelExistenceDirective;
@@ -16,18 +17,9 @@ class ForceDeleteDirective extends ModifyModelExistenceDirective
     {
         return /** @lang GraphQL */ <<<'GRAPHQL'
 """
-Permanently remove one or more soft deleted models by their ID.
-The field must have a single non-null argument that may be a list.
+Permanently remove one or more soft deleted models.
 """
 directive @forceDelete(
-  """
-  DEPRECATED use @globalId, will be removed in v6
-
-  Set to `true` to use global ids for finding the model.
-  If set to `false`, regular non-global ids are used.
-  """
-  globalId: Boolean = false
-
   """
   Specify the class name of the model to use.
   This is only needed when the default model detection does not work.
@@ -37,11 +29,11 @@ directive @forceDelete(
 GRAPHQL;
     }
 
-    protected function find(string $modelClass, $idOrIds)
+    protected function enhanceBuilder(EloquentBuilder $builder): EloquentBuilder
     {
         /** @see \Illuminate\Database\Eloquent\SoftDeletes */
         // @phpstan-ignore-next-line because it involves mixins
-        return $modelClass::withTrashed()->find($idOrIds);
+        return $builder->withTrashed();
     }
 
     protected function modifyExistence(Model $model): bool